refactor: breaking out parts
