+++ app/app.go
@@ -1,7 +1,6 @@
 package app
 
 import (
-	"io"
 	"log"
 	"net/http"
 
@@ -21,27 +20,6 @@ func NewApp(router *mux.Router) App {
 	}
 
 	return app
-}
-
-// Routes holds all registered Routes and universal middleware for the App.
-func (app *App) Routes() {
-	app.router.Use(func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-			writer.Header().Set("Content-Type", "application/json")
-
-			next.ServeHTTP(writer, request)
-		})
-	})
-
-	app.router.HandleFunc("/", app.handleIndex()).Methods(http.MethodGet)
-}
-
-func (app *App) handleIndex() http.HandlerFunc {
-	return func(writer http.ResponseWriter, request *http.Request) {
-		writer.WriteHeader(http.StatusOK)
-
-		_, _ = io.WriteString(writer, "Hello world!")
-	}
 }
 
 // StartServer starts the HTTP server on the specified port. Any errors will be returned on the specified channel.
